Show correct heirloom/scaled items in comments
Default API's GetItemInfo does not provide the correct item level for
heirloom/scaled items (such as some of the new items from the 13th
anniversary world bosses).

LibItemUpgradeInfo performs tooltip scanning to get the correct item
level which I have the simc addon now using to make sure that the
comments for items line up with what is seen in-game.